fts: fix high polling rate after suspend & resume
High polling rate stops working after turning the screen off and
then turning it back on. It seems that the hardware (or firmware)
has it disabled after suspend, but the mode is not updated correctly
in the driver. As a result, after resume, high polling rate appears
to be enabled, but the actual polling rate is stuck at 240Hz max.

Therefore, reset the mode when going into suspend, and re-enable
it after resume.

Test: Touch sampling rate reaches 400Hz+ after suspend & resume

Co-authored-by: EcrosoftXiao <xy19990903@gmail.com>
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>